+++ hibernate-dialect/src/test/java/tech/ydb/hibernate/user/UserRepositoryTest.java
@@ -3,7 +3,6 @@ package tech.ydb.hibernate.user;
 import java.sql.Date;
 import java.sql.Timestamp;
 import java.time.Instant;
-import java.time.LocalDateTime;
 import org.hibernate.cfg.AvailableSettings;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -81,10 +80,6 @@ public class UserRepositoryTest {
 
                     Date date = session
                             .createQuery("select current date", Date.class).getSingleResult();
-
-                    LocalDateTime localDateTime = session
-                            .createQuery("select current time", LocalDateTime.class)
-                            .getSingleResult();
                 }
         ));
     }